AUT-4248: Add utility to find MFAs in a list of MFAMethods
Added `getMfaMethodOrDefaultMfaMethod` as a static method on `MFAMethodsService`. This filters a passed in list of `mfaMethods` based on arguments.

- If an `mfaMethodId` is passed in it will return the method if present, or an empty Optional.
- If `mfaMethodId` is null, it returns the `DEFAULT` MFA method in the list.
- If `methodType` is not null, it will further filter by the type of MFA method.

This will be used in various lambdas to facilitate a lookup of MFA by ID, with a fallback if the ID is not provided. This is so we can add a `mfaMethodId` property to our API request objects so we can target chosen methods, but fallback to the existing behaviour if not provided.

Eventually we could make `mfaMethodId` a mandatory property on our APIs, and not need to do this fallback.