AUT-4248: Verify SMS OTP with identified MFAMethod for sign in
Allow the `/verify-code` API to accept a new `mfaMethodId` attribute, which when provided will use the identified MFA method as the target for verifying the SMS OTP on the sign in journey.

If no ID is sent, it will fallback to the previous behaviour (using the DEFAULT SMS MFAMethod).

If ID is provided but not found for the user we return a 400 Bad Request.