AUT-4574: Use canReceiveEmailAddress for initial lockout check
We are moving lockout business logic to the PermissionDecisionManager. Here we make sure on the first lockout check in the CheckReAuthUserHandler that we use the canReceiveEmailAddress method.

Note that we currently do not support returning all the reauth counts needed for the AUTH_REAUTH_FAILED audit event, hence the TODO reference to AUT-4755. This will be handled in a future commit.